+++ src/main/java/com/vehicle/app/vehicleApi/Repo/VehicleRepo.java
@@ -3,5 +3,8 @@ package com.vehicle.app.vehicleApi.Repo;
 import com.vehicle.app.vehicleApi.Models.Vehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepo extends JpaRepository<Vehicle, Long> {
+
+
+public interface VehicleRepo extends JpaRepository<Vehicle, Integer> {
+
 }
